use cmd server on all platforms
